[esperanto-tools-libs] [SW-3760] Associate the CodeRegistry per device.
This commit create a CodeRegistry instance per Device, to correctly
associate/track the code loaded in each device. Currenty the
Host-Runtime holds a single instance across all devices that is
fuctionally incorrect.

A bug derived from the above is that the CodeRegistry gets destroyed
often after the Device + MemoryAllocator and that causes crashes
where we try to deallocate a CodeBuffer after the MemoryManager of
the Device has been destroyed.

To maintain the existing intreface until we fix all callers the
existing code will fetch the registry of the active device.