[esperanto-tools-libs] [SW-3777] Make DeviceBuffer a fat-pointer as well
Pass the size of the allocation to the DeviceBuffer to enable
doing bounds checks when we are doing memcpy operations.

Fix the memcpy template function to perform a check of the size
of the DeviceBuffer we read/write to.

Computing the direction of the memcpy has been converted to a
constexpr function who result is evaluated at compile time
based on the types of the arguments of the memcpy function.