[device-minion-runtime] FW support for processing PMC configure command.
The master minion sends a broadcast message to all worker minions
with the address of the PMC configuration buffer.

When worker minions receive the message they make the appropriate
S-mode to M-mode syscall.

For the time there is no PMC support for the master shire.
(I need to make sure first how to signal minions in the master shire).