[sw-sysemu] [SW-6543] Skip the timestamp read coherency check for L2Scp
A needed condition for cache coherency on read is that the timestamp for the
read location matches the timestamp for the line on L3, if present.

When writes are done, the timestamps stop matching. Then one then needs to
do evict_va for propagating the timestamp from L1 to higher levels for
restoring cache coherency before the next read.

The L2Scp is an exception to this rule because an L2Scp line can not be
evicted with evict_va to L3, and therefore its timestamp can not be
propagated to L3. However, L2Scp lines can be read without read coherency
issues because the hardware would never evict them, and therefore the
cache coherency condition:
- is impossible to gurantee.
- is redundant

Therefore we skip the cache coherency check before reading from L2Scp
lines.

This patch fixes an issue seen with a Splat test that was failing with
sysemu aborting with a read coherency error message. The test was doing
the following sequence of operations for minion 0:
- store to 0xc0002095
- evict_va 0xc0002080 to L2
- read 0xc0002095, with a false ready coherency issue being reported.

The reported cache coherency issue was due to the L2 cache line
timestamp not matching the one the L3 cache line because of the
aforementioned reasons.