[devicelayer] fix race condition in quick response reception
executor thread was previously submitting the command and on success
adding it's entry to the list of commands sent under current execution
scope. But there was a race condition where response was getting
received before the command entry is saved. So making sure that
if command is sent then it's entry must exist there already.

Signed-off-by: Haseeb Ashraf <Haseeb_Ashraf@mentor.com>