[device-bootloaders] test_framework: align with latest tf_specifications.json
The specifications through the json listings in the bringup
tests have been updated to accomodate new tests that are
coming up so we align the protocol definitions accordingly
here as well. We also now define handler for every index from
the tf_cmd_t and use the unregistered handler wherever the
actual handler implementation isn't found. This helps keeping
track of what needs to be done.

Signed-off-by: Awais Belal <awais_belal@mentor.com>