[et-driver] [SW-8442]: et_vqueue: detect hang-ups using a timer
This implements message counting ability to all the
CQs/SQs and then uses that with a timer to detect
hang-ups on the device side. It is assumed here that
each SQ message will have a corresponding CQ message
which would return in a given time, failing which the
device would be considered in a hang-up.
If the timer isn't running when a SQ push is requested
the timer will be started, each successful CQ pop will
result in resetting/modifying the timer to make sure
it runs atleast once at the end of long running
transactions just to confirm a message is not hung
up on the device side.
The timer matches SQ/CQ message counts when its timeout
is triggered and a mismatch in counts raises an error
message through the error printing interface.

Signed-off-by: Awais Belal <awais_belal@mentor.com>