[et-driver] et_vqueue: increase ET_MSG_TIMEOUT to 300s
The default timeout for the hang-up detection timer was
set to 5 seconds which isn't sufficient for commands such
as DMA that take way longer to respond depending upon
factors such as the size of actual DMA. The largest allowed
DMA is 128MB per node/channel whereas we've 4 such nodes and
a 128MB DMA on average takes around 60s. Keeping this as a
base we update the ET_MSG_TIMEOUT to 300s to accomodate for
such scenarios.

Signed-off-by: Awais Belal <awais_belal@mentor.com>