[et-driver] et_vqueue: fix return codes from et_cqueue_copy_to_user
Returning ENOMEM means the driver was not able to allocate
the required memory which is not true for the cases we fix
here.
When the userspace passes on a smaller buffer than required
for the transaction it is an EINVAL. Also it is a convention
to use EFAULT when copy_to_user fails.

Signed-off-by: Awais Belal <awais_belal@mentor.com>