[sw-sysemu] [BEMU] Track hart and tensor state in BEMU
BEMU now tracks the hart state (nonexistent, unavailable, halted, running,
stalled, etc.) and the tensor unit state. BEMU models the state of the various
tensor FSMs and keeps track of when cooperative tensor loads and tensor
reductions are synchronized. It also models the resource dependences among the
various tensor FSMs and the hart instructions that access the VPU.

Changes include:
- All of SysEmu fcc, msgport, and wfi tracking state and code has been
  removed.
- SysEmu running, waiting, and wake-up on interrupt code has been removed.
- Running, waiting, etc. harts are now tracked in BEMU using a new intrusive
  list implementation that avoids dynamic memory allocation/deallocation.
- Moved cooperative tensor load, tensor reduction, and tensor_wait
  synchronization from SysEmu to BEMU. CoSim currently bypasses the BEMU
  synchronization logic.
- Added the ability to restart an instruction in BEMU. This is necessary to
  model VPU instruction dependences to pending tensor operations, and to model
  the STALL instruction behavior.
- Moved tensor wait synchronization from SysEmu to BEMU. CoSim
  currently bypasses the BEMU synchronization logic.
- Some class/struct names have been renamed from like_this_t to Like_this to
  match the recommended code style.
- Many enum fields have been renamed from LikeThis to like_that to match the
  recommended code style.
- Some if/else/while brackets and indentation has been changed to match the
  recommended code style.
- Some trailing spaces have been removed from the code.
- Move lazy_array code to emu/bemu/support/ and added intrusive::list code
  in the same directory.

A running hart now belongs either in the 'active', 'sleeping', or 'awaking'
list in BEMU.  The active list holds all harts that have actions to performe
(either fetch/execute a new instruction or a background tensor operation).
Sleeping is the contrary of active, and awaking is the transition from
sleeping to active.

The Hart::state variable holds the hart state (nonexistent, unavailable,
active, sleeping). Hart stall conditions are recorded in the Hart::waits flags
(one bit per reason for stalling).