[device-bootloaders] Fix bug due to misuse of SP offset_per_hart
Usually offset_per_hart can be assumed to be start of new entry in trace buffer
but this is not always valid. If space requested by trace entry is more that the
space available till end of trace buffer, wrap around will occur and offset_per_hart
will change. So usage of offset_per_hart before call to Trace_Buffer_Reserve() is
a bug. Fix this issue by modifying bl2_dump_stack_frame() to return pointer to start
of exception entry obtained in result of call to Trace_Buffer_Reserve().

Signed-off-by: Fahad Arslan <fahad_arslan@mentor.com>