[sw-sysemu] [BEMU] Change logging level through instruction triggers
Allow to raise and lower the logging level using instruction triggers to reduce
the amount of logging in cases when only a region of code is interesting. This
simplifies that case for logging the execution of SW models since it uses the
methodology already in place to generate full waves.

Resolves: ARCHSIM-681