[sw-sysemu] BEMU: Model program buffer
The program buffer is modelled as a 4 x 32-bit buffer.
It can hold 4 instructions / 8 compressed instructions.

As far as BEMU is concerned, the instructions in the program buffer
have addresses/PCs in the fixed [0x1800007C0, 0x1800007D0) range.
This allows us to reuse the standard execute/insn_util functions.
Note: 0x1800007C0 corresponds to the ABSCMD address for hart 0.

In esrs:
- HASTATUS1.busy (lower 16 bits) should be read-only
- Reading ABSCMD, AXPROGBUF*, and NXPROGBUF* returns 0
- Writing ABSCMD, AXPROGBUF*, and NXPROGBUF* will update the program buffer
- Writing ABSCMD, and AXPROGBUF* will also enter into the program buffer
- Note: ABSCMD corresponds to bits 0-63 (2 insns) of the program buffer,
  *XPROGBUF0 to bits 64-95, and *XPROGBUF1 to bits 96-127 (1 insn each)
- Writing NXDATA* updates ddata0
- Writing AXDATA* will also enter into the program buffer

In processor:
- Model the program buffer as described above
- Update HASTATUS1 when entering/exiting the program buffer

In insns/system:
- While in debug mode, ebreak will exit the program buffer

In sysemu:
- Fetch/execute instructions while halted and in program buffer