[sw-sysemu] BEMU: Use global address for read bus errors
MemoryRegion::read() may throw a memory_error on a bus error.
This memory error will have the address *local* to the device/region.
In order for other components, such as CoSim, to actually process this
error we need to translate to the *global* address, i.e. the address
seen by the system. This was already done for MemoryRegion::write()s,
but the read()s can also produce a bus error (see SvcProcRegion).