[sw-sysemu] BEMU: Add first draft of VPURF checker
The checker keeps score of write accesses to the VPURF.
On a read we check for RAW issues or applicable workarounds.

Note that this version does not yet support workarounds using fclass.
Since BEMU does not model intpipe timing, taken branches which hide RAW
issues in RTL, will not hide the issue here and be flagged as an error.

See also: FV-271