[sw-sysemu] BEMU: Trigger bus error on illegal fetch/evict
We should trigger a bus error when the source of an evict or flush is
backed by the L2 scratchpad and the destination is L3 or memory.

Note: For evict_va/flush_va we were incorrectly setting tensor_error[7].
Either way BEMU is imprecise here, because we don't model the cache.

Resolves: ARCHSIM-688