[sw-sysemu] Add pause/resume to sysemu interface
The host can now request a `pause()`. This will set a hint, which will
cause the device to sleep for 100 ms every simulation cycle until the
hint is cleared.

The pause hint can be cleared via the `resume()` call, or implicitly when
requesting any MMIO access or raising an interrupt.