[sw-sysemu] BEMU: Add hints to waive read coherency issues
The following hints will be added:
- slti x0,x0,0x603  # Waive read coherency errors
- slti x0,x0,0x604  # Stop waiving errors (the default)

These hints can be used to surround a region of code that has false
sharing issues in order to waive the mem_checker errors, ex:

    flw.ps   0x8005e5e03c..0x8005e5e058
    fscwg.ps 0x8005e5e03c..0x8005e5e058
    flw.ps   0x8005e5e05c..0x8005e5e078 <-- Read Coherency Error

In this example the stale data in L1/L2 would still be up-to-date.
However, since it falls into the same cache line as the following flw.ps
it will be flagged as an error.

See also: ARCHSIM-692