[sw-sysemu] BEMU: Add hints to waive write coherency issues
The following hints will be added:
- slti x0,x0,0x605  # Waive write coherency errors
- slti x0,x0,0x606  # Stop waiving errors (the default)

These hints can be used to surround a region of code that has false
sharing issues in order to waive the mem_checker errors, ex:

    tensor_store 0x8005bb2000..0x8005bb201f
    fscwg.ps     0x8005bb2020..0x8005bb2027 <-- Write Coherency Error

In this example there is no overlap between the tesnor store and the
global scatter. However, the coalescing buffer is dirty for the same
cache line as the following fscwg.ps but for another region of the
cache line. Nevertheless just by being on the same cache line the scatter
will be flagged as an error.

See also: ARCHSIM-697