[device-minion-runtime] [SW-19533] Fixing kernel arguments corruption seen with clang
The inline assembly operands k_param_a{0..3} could be allocated by
the compiler on registers that were that were overwritten by the
inline assembly before assigned to a{0..3}, where the kernel expects
them. This is a bug, and we were lucky enough not hitting it with gcc,
but we were not so lucky with clang.

The assignment now happens earlier, before any potential overwrite,
fixing the issue.

Also the clobber list was empty, when two registers were actually
clobbered, being another source of potential issues. That has not yet
caused us any trouble, but it is also a bug.

A memory location was marked as an output for the inline assembly block,
but the reality is that many memory locations are written, and the
kernel can write anywhere. A broad declaration of intend telling the
compiler that we clobber the "memory" virtual operand seems to better
describe what the inline assembly block and the invoked compute kernel
potentially do.

Finally, there was the possibility that k_entry was allocated on t0
(x5) and was overwritten before consumed. The way of fixing the
potential problem was again to consume k_entry before t0 is produced.