[device-minion-runtime] [SW-19533] Mark ra register as clobbered
The ra register was clobbered but the compiler was not informed. This
mistake does not lead to an error because the function calls other
functions, likely forcing ra to be spilled and restored when the
function returns.

If we did refactor the code and put the inline assembly block into its
own function we would start seeing an issue, unless we apply now this
simple patch to declare the ra clobber.

Conflicts:
	src/WorkerMinion/src/kernel.c