CMakeLists.txt: Build device code in Release mode.
At the moment, non-release build do not work. The culprit seems to
be launchKernel use of C -- it can only work if the functions are
aggressively inlined and there's no access to memory between flushing
of the L1 and re-initializing it.

This will have to be fixed in the proper way, but for now keep the
current code working.